add spec helpers for images and artwork comparisons
